argc and argv commits

+++ 0x0A-argc_argv/3-mul.c
@@ -13,11 +13,11 @@ int main(int argc, char **argv)
 if (argc > 1)
 {
 int i;
-long mul;
+int mul;
 mul = 1;
 for (i = 1; i < argc; i++)
 mul *= atoi(argv[i]);
-printf("%ld\n", mul);
+printf("%d\n", mul);
 }
 else
 {
